update sample loadmore

+++ app/src/main/java/com/lyhoangvinh/simple/ui/base/interfaces/UiRefreshable.kt
@@ -3,5 +3,5 @@ package com.lyhoangvinh.simple.ui.base.interfaces
 interface UiRefreshable : Refreshable {
     fun doneRefresh()
     fun refreshWithUi()
-    fun refreshWithUi(delay: Int)
+    fun refreshWithUi(delay: Long)
 }
